meta/pkg-build/thread: fix test
Add synchronization so that the enumerated interleavings cover
all of the ones that should be possible.

Using `(sync (system-idle-evt))` as a barrier works as long as
the implementation of the library being tested doesn't itself
use `(sync (system-idle-evt))`. That seems like a safe-enough
assumption for a test.

original commit: 59da3bd83e6e7b57a7b1b886339a165d47fad076