meta/pkg-build: add support for "minimal" vs. full VMs
Try building on a minimal VM, first, and then fall back to a more
general VM is that fails. For example, a minimal VM might not include
a C compiler, which can help clarify that a package will work in some
environments and not in others.

original commit: abf76be7f0b9ad3394fa02ae47654e9dda687d11